requirements.txt
aiofiles==22.1.0
aiosqlite==0.19.0
anyio==3.7.1
appnope==0.1.3
arcgis==2.2.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.0
attrs==23.1.0
Babel==2.13.0
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.1.0
cachetools==5.3.1
certifi==2023.7.22
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.0
click==8.1.7
cloudpickle==3.0.0
comm==0.1.4
contourpy==1.1.1
cryptography==41.0.4
cycler==0.12.1
dask==2023.10.0
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
deptry==0.12.0
exceptiongroup==1.1.3
executing==2.0.0
fastjsonschema==2.18.1
fonttools==4.43.1
fqdn==1.5.1
fsspec==2023.9.2
geomet==1.0.0
gssapi==1.8.3
idna==3.4
importlib-metadata==6.8.0
install==1.3.5
ipykernel==6.25.2
ipython==8.16.1
ipython-genutils==0.2.0
ipywidgets==7.8.1
isoduration==20.11.0
jaraco.classes==3.3.0
jedi==0.19.1
Jinja2==3.1.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
jupyter-client==6.1.12
jupyter-events==0.8.0
jupyter-server==1.24.0
jupyter-ydoc==0.2.5
jupyter_core==5.4.0
jupyter_server_fileid==0.9.0
jupyter_server_ydoc==0.8.0
jupyterlab==3.6.6
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.7
jupyterlab_server==2.25.0
keyring==24.2.0
kiwisolver==1.4.5
krb5==0.5.1
locket==1.0.0
lxml==4.9.3
MarkupSafe==2.1.3
matplotlib==3.8.0
matplotlib-inline==0.1.6
mistune==3.0.2
more-itertools==10.1.0
nbclassic==1.0.0
nbclient==0.8.0
nbconvert==7.9.2
nbformat==5.9.2
nest-asyncio==1.5.8
notebook==6.5.6
notebook_shim==0.2.3
numpy==1.26.1
oauthlib==3.2.2
packaging==23.2
pandas==2.1.1
pandocfilters==1.5.0
parso==0.8.3
partd==1.4.1
pathspec==0.11.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.1.0
platformdirs==3.11.0
prometheus-client==0.17.1
prompt-toolkit==3.0.39
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.16.1
pylerc==4.0
pyparsing==3.1.1
pyshp==2.3.1
pyspnego==0.10.2
python-dateutil==2.8.2
python-dotenv==1.0.0
python-json-logger==2.0.7
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==24.0.1
referencing==0.30.2
requests==2.31.0
requests-gssapi==1.2.3
requests-kerberos==0.14.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.10.6
Send2Trash==1.8.2
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
stack-data==0.6.3
terminado==0.17.1
tinycss2==1.2.1
tomli==2.0.1
toolz==0.12.0
tornado==6.3.3
traitlets==5.11.2
truststore==0.8.0
types-python-dateutil==2.8.19.14
tzdata==2023.3
ujson==5.8.0
uri-template==1.3.0
urllib3==2.0.6
wcwidth==0.2.8
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.4
widgetsnbextension==3.6.6
y-py==0.6.2
ypy-websocket==0.8.4
zipp==3.17.0
